+++ src/info-pages/dto/update-info-page.dto.ts
@@ -1,0 +1,4 @@
+import { PartialType } from '@nestjs/swagger';
+import { CreateInfoPageDto } from './create-info-page.dto';
+
+export class UpdateInfoPageDto extends PartialType(CreateInfoPageDto) {}